refactor trans2 code
