+++ src/smb.rs
@@ -7,6 +7,4 @@ pub mod trans;
 pub mod trans2;
 
 pub use self::error::Error;
-pub use self::common::Capabilities;
-pub use self::msg::Msg;
-pub use self::reply::Reply;
+pub use self::common::{Capabilities, DirInfo};
