+++ src/smb/reply.rs
@@ -6,21 +6,21 @@ use crate::win::*;
 use super::Error;
 use super::common::*;
 use super::info::*;
-use super::{trans, trans2};
+use super::trans;
 
 /// Helper struct holding all relevant information of a reply
 /// to be used by create() method.
-pub struct ReplyCtx {
-    info: Info,
+pub(crate) struct ReplyCtx {
+    pub info: Info,
 
-    parameter: Bytes,
+    pub parameter: Bytes,
 
-    data: Bytes,
-    data_offset: usize,
+    pub data: Bytes,
+    pub data_offset: usize,
 }
 
 
-pub trait Reply: Sized {
+pub(crate) trait Reply: Sized {
     const CMD: Cmd;
     const ANDX: bool;
 
@@ -458,51 +458,6 @@ impl<T: trans::SubReply> Reply for Transact<T> {
         Ok(Transact::<T> { subcmd })
     }
 }
-
-
-/// Reply to SMB_COM_TRANSACTION2, see 2.2.4.46.2
-pub struct Transact2<T> {
-    pub subcmd: T,
-}
-
-impl<T: trans2::SubReply> Reply for Transact2<T> {
-    const CMD: Cmd = Cmd::Transact2;
-    const ANDX: bool = false;
-
-    fn create(mut ctx: ReplyCtx) -> Result<Self, Error> {
-        // parameter
-        let total_parameter_count = ctx.parameter.get_u16_le() as usize;
-        let total_data_count = ctx.parameter.get_u16_le() as usize;
-        ctx.parameter.advance(2);   // reserved
-
-        let parameter_count = ctx.parameter.get_u16_le() as usize;
-        let parameter_offset = utils::try_sub(ctx.parameter.get_u16_le().into(), ctx.data_offset)
-            .ok_or(Error::InvalidData)?;
-        ctx.parameter.advance(2);   // ignoring parameter displacement
-
-        let data_count = ctx.parameter.get_u16_le() as usize;
-        let data_offset = utils::try_sub(ctx.parameter.get_u16_le().into(), ctx.data_offset)
-            .ok_or(Error::InvalidData)?;
-        ctx.parameter.advance(2);   // ignoring data displacement
-
-
-        // FIXME we need to support incomplete transact2 replies
-        if parameter_count < total_parameter_count || data_count < total_data_count {
-            return Err(Error::Unsupported("transaction2 reply split to multiple packets".to_owned()));
-        }
-
-        // data
-        let sub_parameter = ctx.data.slice(parameter_offset..parameter_offset+parameter_count);
-        let sub_data = ctx.data.slice(data_offset..data_offset+data_count);
-
-        // create sub-command response
-        let subcmd = T::parse(sub_parameter, sub_data)?;
-
-        Ok(Transact2::<T> { subcmd })
-    }
-}
-
-
 
 
 #[cfg(test)]
