+++ src/smb/trans2/subcmd.rs
@@ -1,0 +1,3 @@
+pub mod findfirst2;
+
+pub use self::findfirst2::FindFirst2;
